WIP REPL (fix mem leaks)

+++ compiler/lexer/lexinput.h
@@ -4,7 +4,7 @@
 
 struct LEXINPUT;
 struct LEXINPUT *lexinput_new_file(FILE *lp);
-struct LEXINPUT *lexinput_new_bb(char *buf, int len);
+struct LEXINPUT *lexinput_new_bb(char *buf, size_t len);
 int lxgetc(struct LEXINPUT *lp);
 int lxtell(struct LEXINPUT *lp);
 int lxseek(struct LEXINPUT *lp, int w, int cmd);
