+++ yasl.h
@@ -2,6 +2,7 @@
 
 #include "yasl_error.h"
 #include "inttypes.h"
+#include <stdlib.h>
 
 struct YASL_State;
 struct YASL_Object;
@@ -13,7 +14,7 @@ struct YASL_Object;
 struct YASL_State *YASL_newstate(char *filename);
 struct YASL_State *YASL_newstate_bb(char *buf, int len);
 
-void YASL_resetstate_bb(struct YASL_State *S, char *buf, int len);
+void YASL_resetstate_bb(struct YASL_State *S, char *buf, size_t len);
 
 /**
  * deletes the given YASL_State.
